requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Flask-Mail==0.9.1
Werkzeug==2.2.2
python-dotenv==1.0.0
cryptography==40.0.2
email-validator==2.0.0.post2
PyJWT==2.6.0